clean up schemas
